:add for float

+++ Learngit/Sum.py
@@ -1,4 +1,4 @@
-a = int(input())
-b = int(input())
+a = float(input())
+b = float(input())
 sum = (a+b)
 print(sum)
